moved mock socket and wsstream
