+++ server/src/managers/mod.rs
@@ -2,6 +2,6 @@ pub mod account_manager;
 mod client_presence_manager;
 pub mod key_manager;
 pub mod messages_manager;
-mod mock_db;
+mod mock_helper;
 pub mod state;
 pub mod websocket;
