+++ server/src/managers/websocket/mod.rs
@@ -1,3 +1,4 @@
 pub mod connection;
 mod net_helper;
 pub mod websocket_manager;
+pub mod wsstream;
